refactor(perf): split src/cli/plan.ts — extract helpers and decompose command
plan.ts (807→415 lines) split into plan-helpers.ts (128 lines) for pure
utility functions and plan-decompose.ts (270 lines) for planDecomposeCommand
and runReplanLoop. Uses lazy dynamic import in _planDeps.planDecompose to
avoid circular dependency. Barrel index.ts updated to re-export from new files.
